requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.12.3
cffi==1.17.1
channels==4.2.2
constantly==23.10.4
crispy-bootstrap5==2025.6
cryptography==45.0.5
daphne==4.1.0
Django==5.2.4
django-ckeditor-5==0.2.18
django-crispy-forms==2.4
django-extensions==4.1
django-jazzmin==3.0.1
django-taggit==6.1.0
django-turnstile==0.1.2
djangorestframework==3.16.0
google-generativeai==0.7.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
packaging==25.0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-dotenv==1.1.1
requests==2.32.3
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.9.0
zope.interface==7.2
coverage==7.9.2
ruff==0.12.3
Cython==3.1.2
django-widget-tweaks==1.5.0
psycopg2-binary==2.9.9
python-json-logger==2.0.7
django-prometheus==2.3.1
django-ratelimit==4.1.0
django-axes==6.3.0
python-magic==0.4.27
django-ipware==6.0.0
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
flower==2.0.1